accualy iz duble
Distance formula now returns a double.

+++ src/model/map/DiscreteMap.java
@@ -33,7 +33,7 @@ public abstract class DiscreteMap {
     public abstract Coordinate2D getCoordinateByLocation( Location l );
     public abstract Location getLocationByMapObject( MapObject m );
     public abstract Location getPreferredTeleportLocation();
-    public abstract int getDistance( MapObject m1, MapObject m2 );
+    public abstract double getDistance( MapObject m1, MapObject m2 );
 
     public String getName() {
         return name;
