+++ src/model/map/GameMap.java
@@ -169,14 +169,14 @@ public class GameMap extends DiscreteMap {
     }
 
     @Override
-    public int getDistance(MapObject m1, MapObject m2) {
+    public double getDistance(MapObject m1, MapObject m2) {
         HexCoordinate h1 = getMapObjectCoordinate( m1 );
         HexCoordinate h2 = getMapObjectCoordinate( m2 );
 
         if (h1.getX() == -1 || h1.getY() == -1 || h2.getX() == -1 || h2.getY() == -1)
             return -1;
 
-        return (int) Math.sqrt( Math.pow( h1.getX() - h2.getX(), 2) + Math.pow( h1.getY() - h2.getY(), 2) );
+        return Math.sqrt( Math.pow( h1.getX() - h2.getX(), 2) + Math.pow( h1.getY() - h2.getY(), 2) );
     }
 
     public void performEffect(Effect effect, InfluenceSet influence) {
